requirements.txt
google-cloud-storage
google-cloud-aiplatform
PyPDF2
python-dotenv
langchain
google-cloud-storage
google-cloud-aiplatform
langchain-community
python-dotenv
pypdf
langchain
chromadb # Vector storage
pytest
boto3
langchain-google-vertexai
langchain_google_community
gcs_utils
streamlit


